+++ packages/mango-ui/src/components/Button/index.tsx
@@ -31,5 +31,3 @@ export const Button: React.FC<ButtonProps> = ({
     {children}
   </button>
 );
-
-export default Button;