+++ packages/mango-ui/src/components/Header/index.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, cleanup } from '@testing-library/react';
-import Header from './index';
+import { Header } from './index';
 
 describe('Header Component', () => {
   beforeEach(cleanup);