+++ packages/mango-ui/src/components/Input/index.tsx
@@ -30,5 +30,3 @@ export const Input: React.FC<InputProps> = ({
     />
   );
 };
-
-export default Input;