+++ packages/mango-ui/src/index.ts
@@ -1,27 +1,7 @@
-import Button from './components/Button';
-
-import Editor from './components/Editor';
-
-import Input from './components/Input';
-
-import Radio from './components/Radio';
-
-import Header from './components/Header';
-
-import Animation from './components/Animation';
-
 export * from './components/Button';
 export * from './components/Editor';
 export * from './components/Input';
 export * from './components/Radio';
 export * from './components/Header';
 export * from './components/Animation';
-
-export default {
-  Button,
-  Editor,
-  Input,
-  Radio,
-  Header,
-  Animation,
-};
+export * from './components/Menu';