+++ packages/mango-ui/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.40](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/mango-ui@0.0.39...@mango-solution/mango-ui@0.0.40) (2021-06-27)
+
+**Note:** Version bump only for package @mango-solution/mango-ui
+
+
+
+
+
+## [0.0.39](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/mango-ui@0.0.38...@mango-solution/mango-ui@0.0.39) (2021-06-27)
+
+**Note:** Version bump only for package @mango-solution/mango-ui
+
+
+
+
+
+## [0.0.38](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/mango-ui@0.0.37...@mango-solution/mango-ui@0.0.38) (2021-06-27)
+
+**Note:** Version bump only for package @mango-solution/mango-ui
+
+
+
+
+
+## [0.0.37](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/mango-ui@0.0.36...@mango-solution/mango-ui@0.0.37) (2021-06-27)
+
+**Note:** Version bump only for package @mango-solution/mango-ui
+
+
+
+
+
 ## [0.0.36](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/mango-ui@0.0.35...@mango-solution/mango-ui@0.0.36) (2021-06-24)
 
 **Note:** Version bump only for package @mango-solution/mango-ui