+++ packages/stories/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.36](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/stories@0.0.35...@mango-solution/stories@0.0.36) (2021-06-27)
+
+**Note:** Version bump only for package @mango-solution/stories
+
+
+
+
+
+## [0.0.35](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/stories@0.0.34...@mango-solution/stories@0.0.35) (2021-06-27)
+
+**Note:** Version bump only for package @mango-solution/stories
+
+
+
+
+
+## [0.0.34](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/stories@0.0.33...@mango-solution/stories@0.0.34) (2021-06-27)
+
+**Note:** Version bump only for package @mango-solution/stories
+
+
+
+
+
+## [0.0.33](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/stories@0.0.32...@mango-solution/stories@0.0.33) (2021-06-27)
+
+**Note:** Version bump only for package @mango-solution/stories
+
+
+
+
+
 ## [0.0.32](https://github.com/MangoYellowH/mango-toolkit/compare/@mango-solution/stories@0.0.31...@mango-solution/stories@0.0.32) (2021-06-24)
 
 **Note:** Version bump only for package @mango-solution/stories